Data.Text.Encoding.Base64

Description

This module contains the combinators implementing the RFC 4648 specification for the Base64 encoding including unpadded and lenient variants

Documentation

encodeBase64 :: Text -> Text

Encode Text in base64 with padding.

See: RFC-4648 section 5

decodeBase64 :: Text -> Either Text Text

Decode a padded base64 encoded Text value

See: RFC-4648 section 4

encodeBase64Unpadded :: Text -> Text

Encode a Text in base64 without padding.

Note: in some circumstances, the use of padding ("=") in base-encoded data is not required or used. If you are absolutely sure the length of your input data is divisible by 3, this function will be the same as encodeBase64 with padding. However, if not, you may see garbage appended to output in the form of "NUL".

Only call unpadded variants when you can make assumptions about the length of your input data.

See: RFC-4648 section 3.2

decodeBase64Unpadded :: Text -> Either Text Text

Decode an unpadded base64 encoded Text

See: RFC-4648 section 4