Data.Text.Encoding.Base64.Lens

Description

This module contains the HasBase64 and HasBase64Unpadded instances for Text, which defined to be the collection of Prisms defining the RFC 4648 specification for the padded and unpadded Base64 encoding format.

These typeclasses are re-exported for convenience

Documentation

class HasBase64 s where

If a particular type s has a base64 representation for any of its focii, this class provides the optical interface for satisfying the padded base64 spec in RFC 4648

Associated Types

type Base64 s

Methods

_Base64 :: Prism' s (Base64 s)

A prism into a base64-encoded focus of some type

Examples:

>>> _Base64 @Text # "Sun"
"UV3u"

>>> _Base64 @Text # "<<???>>"
"PDw/Pz8+Pg=="

>>> "PDw/Pz8+Pg==" ^. _Base64
"<<???>>"

>>> "PDw/Pz8+Pg==" ^? _Base64
Just "<<???>>"

>>> "PDw_Pz8-Pg==" ^. _Base64
""

>>> "PDw_Pz8-Pg==" ^? _Base64
Nothing

_Base64Url :: Prism' s (Base64 s)

A prism into the base64url-encoded focus of some type

Examples:

>>> _Base64Url @Text # "Sun"
"UV3u"

>>> _Base64Url @Text # "<<???>>"
"PDw_Pz8-Pg=="

>>> "PDw_Pz8-Pg==" ^. _Base64Url
"<<???>>"

>>> "PDw_Pz8-Pg==" ^? _Base64Url
Just "<<???>>"

>>> "PDw/Pz8+Pg==" ^. _Base64Url
""

>>> "PDw/Pz8+Pg==" ^? _Base64Url
Nothing

class HasBase64Unpadded s where

If a particular type a has an unpadded base64 representation for any of its focii, this class provides the optical interface for satisfying the unpadded base64 spec in RFC 4648

Associated Types

type Base64Nopad s

Methods

_Base64Unpadded :: Prism' s (Base64Nopad s)

A prism into the unpadded base64-encoded focus of some type

_Base64UrlUnpadded :: Prism' s (Base64Nopad s)

A prism into the unpadded base64url-encoded focus of some type

Orphan instances

HasBase64Unpadded Text

Instance details

Associated Types

type Base64Nopad Text :: Type

Methods

_Base64Unpadded :: Prism' Text (Base64Nopad Text)

_Base64UrlUnpadded :: Prism' Text (Base64Nopad Text)

HasBase64 Text

Instance details

Associated Types

type Base64 Text :: Type

Methods

_Base64 :: Prism' Text (Base64 Text)

_Base64Url :: Prism' Text (Base64 Text)